Replaced all size_t with int for problem 3 as it only overly complicated the code and also made it more likely for bugs to crop up (I spotted a couple, like casting the result of a size_t - 1 to an int, when it should have been size_t to int followed by -1. That could've underflowed if it happened to start empty.

+++ 2023/Day_03/part1.cpp
@@ -9,8 +9,8 @@
 const char* INPUT_FILE = "input.txt";
 
 struct enginePart {
-    size_t row{0};
-    size_t col{0};
+    int row{0};
+    int col{0};
     int value{0};
     int length{0};
     bool found{false};
@@ -31,8 +31,8 @@ int main() {
 
     // find all groups of digits and extract the integer value,
     // and store its location for easy processing in the next step
-    for (size_t row = 0; row < engineBoard.size(); row++) {
-        for (size_t col = 0; col < engineBoard[row].size(); col++) {
+    for (int row = 0; row < int(engineBoard.size()); row++) {
+        for (int col = 0; col < int(engineBoard[row].size()); col++) {
             if (!std::isdigit(engineBoard[row][col])) {
                 continue;
             }
@@ -51,7 +51,7 @@ int main() {
             }
 
             // if we can scan the next character, increase the bounds
-            if (part.col + part.length + 1 < engineBoard[row].size()) {
+            if (part.col + part.length + 1 < int(engineBoard[row].size())) {
                 part.length++;
             }
         }
@@ -62,8 +62,8 @@ int main() {
     // process all of the numbers we found by scanning for symbols
     // within 1 character of it. std::max and std::min clamp our rows
     for (auto& part : parts) {
-        for (size_t rowOffset = std::max(int(part.row)-1, 0); !part.found && int(rowOffset) <= std::min(int(part.row)+1, int(engineBoard.size()-1)); rowOffset++) {
-            for (size_t colOffset = 0; int(colOffset) < part.length; colOffset++) {
+        for (int rowOffset = std::max(part.row-1, 0); !part.found && rowOffset <= std::min(part.row+1, int(engineBoard.size())-1); rowOffset++) {
+            for (int colOffset = 0; colOffset < part.length; colOffset++) {
                 char symbol = engineBoard[rowOffset][part.col+colOffset];
 
                 if (!std::isdigit(symbol) && symbol != '.') {
